+++ db/migrate/20151026161928_add_recruiting_center_to_sites.rb
@@ -1,0 +1,6 @@
+class AddRecruitingCenterToSites < ActiveRecord::Migration
+  def change
+    add_column :sites, :recruiting_center, :boolean, null: false, default: false
+    add_index :sites, :recruiting_center
+  end
+end