+++ db/migrate/20151026161941_add_coordinating_center_to_sites.rb
@@ -1,0 +1,6 @@
+class AddCoordinatingCenterToSites < ActiveRecord::Migration
+  def change
+    add_column :sites, :coordinating_center, :boolean, null: false, default: false
+    add_index :sites, :coordinating_center
+  end
+end