+++ db/migrate/20151026170014_add_contact_to_sites.rb
@@ -1,0 +1,5 @@
+class AddContactToSites < ActiveRecord::Migration
+  def change
+    add_column :sites, :contact, :string
+  end
+end